+++ src/main/java/racing/dto/RacingResult.java
@@ -1,11 +1,14 @@
-package racing.model;
+package racing.dto;
 
-public class Result {
+import racing.model.Name;
+import racing.model.Position;
+
+public class RacingResult {
 
 	private final int position;
 	private final String name;
 
-	public Result(Position position, Name name) {
+	public RacingResult(Position position, Name name) {
 		this.position = position.getPosition();
 		this.name = name.getName();
 	}
