+++ src/test/java/racing/dto/GameRequestTest.java
@@ -1,4 +1,4 @@
-package racing.view;
+package racing.dto;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
